Revert "Feat/single accordion component"
