+++ src/components/accordion/accordion-DOCS.md
@@ -1,0 +1,9 @@
+# Accordion
+
+Accordion component.
+
+```js
+import { AccordionPrimary, AccordionSecondary, AccordionGroup, AccordionProps } from '@panenco/ui';
+```
+
+<!-- STORY -->
