--- src/components/accordion/accordionItem/types.ts
@@ -1,1 +1,0 @@
-export type AccordionVariant = 'outlined' | 'text';
