+++ src/components/accordion/index.tsx
@@ -1,2 +1,3 @@
-export * from './accordionItem';
+export * from './primary';
+export * from './secondary';
 export * from './group';
