+++ src/components/accordion/useAccordion.tsx
@@ -5,7 +5,7 @@ import { useAccordionContext } from './group';
 export interface UseAccordionProps {
   isOpen?: boolean;
   onClick?: (e?: MouseEvent) => void;
-  ref?: React.RefObject<HTMLDivElement>;
+  ref?: any;
 }
 
 export interface UseAccordinReturn {
@@ -15,7 +15,7 @@ export interface UseAccordinReturn {
 }
 
 export const useAccordion = (props: UseAccordionProps): UseAccordinReturn => {
-  const { isOpen: isOpened = false, onClick, ref } = props;
+  const { isOpen: isOpened, onClick, ref } = props;
   const [isOpen, setOpen] = React.useState(isOpened);
   const innerRef = React.useRef();
   const combinedRef = useCombinedRefs(innerRef, ref);
