+++ stories/accordion/types.ts
@@ -1,0 +1,10 @@
+// export interface AccordionProps extends React.HTMLAttributes<HTMLDivElement> {
+//   iconLeft?: boolean | string;
+//   iconRight?: boolean | string;
+//   icons?: Array<string>;
+//   isOpen?: boolean;
+//   onClick?: any;
+//   onIconClick?: any;
+//   shouldRotateIcon?: boolean;
+//   title?: string;
+// }
