--- stories/accordionGroup/readme.md
@@ -1,1 +1,0 @@
-`AccordionGroup` is a wrapper component that controls the open state automatically to keep only one `Accordion` item open at a time.
